Restructure grant module
Modernize ITs using test containers
