+++ pass-grant-loader/entrypoint.sh
