--- pass-grant-loader/pass-grant-data/src/main/resources/logback.xml
@@ -1,32 +1,0 @@
-<!--
-  ~ Copyright 2023 Johns Hopkins University
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<configuration>
-  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-    <encoder>
-      <pattern>%d{HH:mm:ss.SSS} [%10.30thread] %-5level [%15.-30C{0}] - %msg%n</pattern>
-    </encoder>
-  </appender>
-  <root level="WARN">
-    <appender-ref ref="STDOUT"/>
-  </root>
-  <logger name="org.eclipse.pass" additivity="false">
-    <appender-ref ref="STDOUT"/>
-  </logger>
-  <logger name="org.eclipse.pass.support.grant.data.DifferenceLogger" additivity="false" level="INFO">
-    <appender-ref ref="STDOUT"/>
-  </logger>
-</configuration>
