--- pass-grant-loader/pass-grant-data/src/test/resources/policy.properties
@@ -1,2 +1,0 @@
-moo:policy/moo/policy/policy1
-baa:policy/baa/policy/policy2
