+++ pass-grant-loader/src/test/java/org/eclipse/pass/support/grant/AbstractIntegrationTest.java
@@ -1,0 +1,49 @@
+package org.eclipse.pass.support.grant;
+
+import java.io.FileReader;
+
+import org.apache.maven.model.Model;
+import org.apache.maven.model.io.xpp3.MavenXpp3Reader;
+import org.eclipse.pass.support.client.PassClient;
+import org.junit.jupiter.api.BeforeEach;
+import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.containers.wait.strategy.Wait;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+import org.testcontainers.utility.DockerImageName;
+
+@Testcontainers
+public abstract class AbstractIntegrationTest {
+
+    static {
+        MavenXpp3Reader reader = new MavenXpp3Reader();
+        try {
+            Model model = reader.read(new FileReader("pom.xml"));
+            String version = model.getParent().getVersion();
+            PASS_CORE_IMG = DockerImageName.parse("ghcr.io/eclipse-pass/pass-core-main:" + version);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private static final DockerImageName PASS_CORE_IMG;
+
+    @Container
+    protected static final GenericContainer<?> PASS_CORE_CONTAINER = new GenericContainer<>(PASS_CORE_IMG)
+        .withEnv("PASS_CORE_BASE_URL", "http://localhost:8080")
+        .withEnv("PASS_CORE_BACKEND_USER", "backend")
+        .withEnv("PASS_CORE_BACKEND_PASSWORD", "backend")
+        .waitingFor(Wait.forHttp("/data/grant").forStatusCode(401))
+        .withExposedPorts(8080);
+
+    protected PassClient passClient;
+
+    @BeforeEach
+    void init() {
+        String clientUrl = "http://localhost:" + PASS_CORE_CONTAINER.getMappedPort(8080);
+        System.setProperty("pass.core.url", clientUrl);
+        System.setProperty("pass.core.user", "backend");
+        System.setProperty("pass.core.password", "backend");
+        passClient = PassClient.newInstance();
+    }
+}
